(audio_mixer) Hold the streaming voice's resample ratio as a double
Both resampler APIs take the ratio as a double: struct resampler_data
and struct resampler_data_int16 both declare it that way, and
sinc_resampler_int16_init takes a double bandwidth_mod. The s16
streaming path computed its ratio as a double and then stored it in a
float field on the voice, so the value was narrowed on the way in and
widened again on the way out for nothing.

The decoded s16 path never did this - one_shot_resample_s16 keeps its
double and hands it straight to the resampler. That difference, and not
anything about running two resamplers, is what the residual between a
decoded and a streaming s16 voice actually was.

With the field held as a double the two agree exactly. Across the
corpus every off-rate s16 comparison that previously differed now
matches byte for byte: 26 cases that ranged up to 57 differing samples
and 30 LSB are all zero, none regressed. The earlier reading of that
residual as inherent to two resamplers was wrong.

The float path is unaffected. It computes a float ratio and widening a
float to a double is exact, so every float comparison is unchanged.

+++ libretro-common/audio/audio_mixer.c
@@ -149,7 +149,12 @@ struct audio_mixer_voice
          unsigned    samples;
          unsigned    buf_samples;
          unsigned    channels;    /* source channels; mono is duplicated */
-         float       ratio;
+         /* Both resampler APIs take the ratio as a double, and the s16
+          * path computes it as one. Holding it as a float here meant
+          * that value was narrowed on the way in and widened again on
+          * the way out, losing precision for nothing. The float path
+          * computes a float and widens exactly, so it is unaffected. */
+         double      ratio;
          /* s16 pipeline (parallel; used when voice->is_s16) */
          int16_t    *buffer_s16;
          void       *resampler_int16;
@@ -1383,7 +1388,7 @@ static bool audio_mixer_play_stream_s16(
    voice->types.stream.resampler_int16 = resamp_i16;
    voice->types.stream.buffer_s16      = (int16_t*)sbuf;
    voice->types.stream.buf_samples     = samples;
-   voice->types.stream.ratio           = (float)ratio;
+   voice->types.stream.ratio           = ratio;
    voice->types.stream.stream          = xfer;
    voice->types.stream.position        = 0;
    voice->types.stream.samples         = 0;
